pyproject.toml
[project]
name = "gsc-cli"
version = "0.1.0"
description = "Agent-first CLI for Google Search Console SEO analytics"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
keywords = ["seo", "google-search-console", "cli", "analytics", "search"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
]
dependencies = [
    "typer>=0.9.0",
    "google-api-python-client>=2.100.0",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.1",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-mock", "ruff"]

[project.urls]
Repository = "https://github.com/demegire/gsc-cli"
Issues = "https://github.com/demegire/gsc-cli/issues"

[project.scripts]
gsc = "gsc_cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gsc_cli"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.lint.per-file-ignores]
# CLI option declarations are naturally long
"src/gsc_cli/query/commands.py" = ["E501"]
"src/gsc_cli/inspect/commands.py" = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
